move geoObstacle from reference frame into spatialmath
