+++ spatialmath/geo_obstacle.go
@@ -1,20 +1,18 @@
-package referenceframe
+package spatialmath
 
 import (
 	geo "github.com/kellydunn/golang-geo"
 	commonpb "go.viam.com/api/common/v1"
-
-	"go.viam.com/rdk/spatialmath"
 )
 
 // GeoObstacle is a struct to store the location and geometric structure of an obstacle in a geospatial environment.
 type GeoObstacle struct {
 	location   *geo.Point
-	geometries []spatialmath.Geometry
+	geometries []Geometry
 }
 
 // NewGeoObstacle constructs a GeoObstacle from a geo.Point and a slice of Geometries.
-func NewGeoObstacle(loc *geo.Point, geom []spatialmath.Geometry) *GeoObstacle {
+func NewGeoObstacle(loc *geo.Point, geom []Geometry) *GeoObstacle {
 	return &GeoObstacle{
 		location:   loc,
 		geometries: geom,
@@ -27,7 +25,7 @@ func (gob *GeoObstacle) Location() *geo.Point {
 }
 
 // Geometries returns the geometries which comprise structure of the GeoObstacle.
-func (gob *GeoObstacle) Geometries() []spatialmath.Geometry {
+func (gob *GeoObstacle) Geometries() []Geometry {
 	return gob.geometries
 }
 
@@ -46,9 +44,9 @@ func GeoObstacleToProtobuf(geoObst *GeoObstacle) (*commonpb.GeoObstacle, error) 
 // GeoObstacleFromProtobuf takes a Protobuf representation of a GeoObstacle and converts back into a Go struct.
 func GeoObstacleFromProtobuf(protoGeoObst *commonpb.GeoObstacle) (*GeoObstacle, error) {
 	convPoint := geo.NewPoint(protoGeoObst.GetLocation().GetLatitude(), protoGeoObst.GetLocation().GetLongitude())
-	convGeoms := []spatialmath.Geometry{}
+	convGeoms := []Geometry{}
 	for _, protoGeom := range protoGeoObst.GetGeometries() {
-		newGeom, err := spatialmath.NewGeometryFromProto(protoGeom)
+		newGeom, err := NewGeometryFromProto(protoGeom)
 		if err != nil {
 			return nil, err
 		}
